Added most of the DB stuff
Added DB files, scripts, and necessary files for sqlite to work.
