+++ Assets/Prefabs/Score.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 59839262aaa08314e8f541fc30c38212
-timeCreated: 1510815539
+guid: 68d8756e824aed0479ff1325190c6272
+timeCreated: 1511140008
 licenseType: Free
 NativeFormatImporter:
   mainObjectFileID: 100100000
