+++ Assets/Scenes/LeaderBoard.unity.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 9796e95fc05574b4bb3e0091ea1d459b
-folderAsset: yes
-timeCreated: 1510793180
+guid: c1a23fcc95066c14f9dd596c923b26c2
+timeCreated: 1510991389
 licenseType: Free
 DefaultImporter:
   userData:
