+++ Assets/Scripts/LeaderBoard.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 088bc8c20b9d3554882a2f73c72548c2
-timeCreated: 1510822091
+guid: 8391c4f84587c2b4ab05af8c298e36b8
+folderAsset: yes
+timeCreated: 1511127766
 licenseType: Free
 DefaultImporter:
   userData:
