+++ Assets/Scripts/LeaderBoard/HighScore.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 674e69b1f3f717d4ebeb383400dcdae2
-timeCreated: 1510793180
+guid: 69fc820edfcc3b14b886cb7d9d663df1
+timeCreated: 1511138528
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
