+++ Assets/Scripts/LeaderBoard/HighScoreScript.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4756aea970ccd3e44b76db62f63d9e63
-timeCreated: 1510793180
+guid: 134d83ff3bc7baf42bd1ae5955c90651
+timeCreated: 1511149797
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
