+++ Assets/Scripts/LeaderBoard/LeadBoardManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: dd01ee6a6dd13fc42a7e34a46de7a8b4
-timeCreated: 1510793180
+guid: 2ff188d301bffa14d8aaf4182c6daa2a
+timeCreated: 1511127784
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
